+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 # unreleased
 * Fix issue where extension code was running on subdomains by accident (#26)
+* Fix issue where an error may occur on Scryfall domains that do not contain the Scryfall deckbuilder globals
 * EDHRec Suggestions
   * Add section selector (#23)